Add e2e key package claim test
